+++ src/database/storage.ts
@@ -1007,8 +1007,9 @@ export class _NanoSQLStorage {
                     }
                     if (intersect(["trie", "idx", "idx()", "trie()"], model.props)) {
                         hasIDX = true;
+                        const isNumber = ["number", "float", "int"].indexOf(model.type) !== -1;
                         dataModels["_" + table + "_idx_" + model.key] = [
-                            { key: "id", type: ["number", "float", "int"].indexOf(model.type) !== -1 ? model.type : "string", props: ["pk"] },
+                            { key: "id", type: isNumber ? model.type : "string", props: ["pk"] },
                             { key: "rows", type: "any[]" }
                         ];
                     }
@@ -1016,15 +1017,15 @@ export class _NanoSQLStorage {
                         if (prop.indexOf("search(") !== -1) {
                             hasSearch = true;
                             dataModels["_" + table + "_search_" + model.key] = [
-                                { key: "wrd", type: "string", props: ["pk"] },
+                                { key: "wrd", type: "string", props: ["pk", "ns()"] },
                                 { key: "rows", type: "any[]" }
                             ];
                             dataModels["_" + table + "_search_fuzzy_" + model.key] = [
-                                { key: "wrd", type: "string", props: ["pk"] },
+                                { key: "wrd", type: "string", props: ["pk", "ns()"] },
                                 { key: "rows", type: "any[]" }
                             ];
                             dataModels["_" + table + "_search_tokens_" + model.key] = [
-                                { key: "id", type: pkType, props: ["pk"] },
+                                { key: "id", type: pkType, props: ["pk", "ns()"] },
                                 { key: "hash", type: "string" },
                                 { key: "tokens", type: "any[]" }
                             ];